[Engineering] Share the Postgres harness via conftest, not by importing a fixture
The pre-push suite went green (2575 passed, testsfailed=0, exitstatus=OK) and
then died in TEARDOWN, with two Docker npipe read timeouts and three orphaned
postgres:16-alpine containers left running. The cause was in the previous
commit, not in Docker.

`test_expand_contract.py` reached for the round-trip harness with

    from tests.test_migrations.test_roundtrip import roundtrip_db_url

which looks like sharing and is not. pytest registers an imported fixture as a
SEPARATE FixtureDef in the importing module, so `scope="session"` caches once
per module rather than once per session. `pytest --setup-plan` says it plainly:

    SETUP    S roundtrip_db_url      <- for test_expand_contract.py
    SETUP    S roundtrip_db_url      <- for test_roundtrip.py
    TEARDOWN S roundtrip_db_url
    TEARDOWN S roundtrip_db_url

Two setups meant two Postgres containers per run, each with its own ryuk and
its own teardown -- and on Windows the second teardown timed out against the
Docker named pipe and failed an otherwise passing suite.

Moved the harness (PROJECT_ROOT, _docker_available, _no_postgres, _run_alembic,
roundtrip_db_url) into tests/test_migrations/conftest.py, where a fixture is
discovered by every module in the directory as ONE FixtureDef. Neither test
module imports from the other any more, and the `noqa: F401` fixture re-export
and its matching `noqa: F811` are gone with it.

Measured, not assumed:

  before   SETUP S x2 / TEARDOWN S x2   78 passed in 83.11s
  after    SETUP S x1 / TEARDOWN S x1   78 passed in 23.56s

The 3.5x is the second container's startup disappearing, which is also the
independent confirmation that the duplicate was real.

Guard re-proven after the refactor, since this touched its imports and its skip
path: with a temporary op.drop_column("api_keys", "name") in the chain it still
fails on exactly the right assertion (1 failed, 6 passed), and passes without.

Worth noting for anyone reading later: this was a genuine bug in test
infrastructure that no test asserts against, and it only surfaced because the
teardown happened to time out. `--setup-plan` is the tool that makes it visible.
